光线追踪渲染器 v1.01 bug修复，离屏渲染

+++ src/Ui/BaseOpenGLWidget.h
@@ -54,7 +54,7 @@ protected:
 
     float deltaTime; //渲染每帧间隔时间
     float lastFrameTime; //上一帧时间
-    int fps; //帧数
+    float fps; //帧数
  
     QSharedPointer<SceneManager> sceneManager; //场景管理器
